src/requirements.txt
click==8.1.7
Flask==2.2.5
gunicorn==23.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==0.21.1
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
